Cleaned up App.js and moved form state to form components. Moved yup testing to their own folder. Sign-up and Sign-in both tested and working.
